src/main/java/org/cost/game/GameController.java
package org.cost.game;

import lombok.*;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

@RestController("/game")
public class GameController {

    private GameRepository gameRepository;

    @Autowired
    public GameController(GameRepository gameRepository) {
        this.gameRepository = gameRepository;
    }

    @RequestMapping(path = "/game", method = RequestMethod.POST)
    public ResponseEntity createGame(@RequestBody CreateGameRequest createGameRequest) {
        if (!gameRepository.exists(createGameRequest.getGameName())) {
            gameRepository.save(new Game(createGameRequest.getGameName()));
            return new ResponseEntity(HttpStatus.OK);
        }
        return new ResponseEntity(HttpStatus.CONFLICT);
    }

    @Builder
    @AllArgsConstructor
    @Getter
    @Setter
    public static class CreateGameRequest {
        private String gameName;
    }
}
